+++ tmsm-woocommerce-booking-thalasso.php
@@ -16,7 +16,7 @@
  * Plugin Name:       TMSM WooCommerce Booking Thalasso
  * Plugin URI:        https://github.com/thermesmarins/tmsm-woocommerce-booking-thalasso
  * Description:       Booking Thalasso
- * Version:           1.1.6
+ * Version:           1.1.7
  * Author:            Nicolas Mollet
  * Author URI:        https://github.com/nicomollet
  * License:           GPL-2.0+
@@ -38,7 +38,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Start at version 1.0.0 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define( 'TMSM_WOOCOMMERCE_BOOKING_THALASSO_VERSION', '1.1.6' );
+define( 'TMSM_WOOCOMMERCE_BOOKING_THALASSO_VERSION', '1.1.7' );
 
 /**
  * The code that runs during plugin activation.