+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 * **New** - Elementor Dynamic Tag: Resaweb URL No Accommodation
 * **New** - Elementor Dynamic Tag: Package Price Explanation
 * **Fix** - Order package/accommodation/spatreatment/discovery by order_menu
+* **Tweak** - Allow comments for Package CPT
+* **New** - Accommodation CPT: redirect to resaweb_url
 
 ### 1.1.6: September 11th, 2018
 * **New** - Order package/accommodation/spatreatment/discovery by order_menu