+++ CHANGELOG.md
@@ -1,9 +1,10 @@
+### 1.1.7: September 20th, 2018
 * **New** - Elementor Dynamic Tag: Resaweb URL No Accommodation
 * **New** - Elementor Dynamic Tag: Package Price Explanation
+* **New** - Body class for CPT without urls, to allow hiding parts of the single page
+* **New** - Accommodation CPT: redirect to resaweb_url
 * **Fix** - Order package/accommodation/spatreatment/discovery by order_menu
 * **Tweak** - Allow comments for Package CPT
-* **New** - Accommodation CPT: redirect to resaweb_url
-* **New** - Body class for CPT without urls, to allow hiding parts of the single page
 
 ### 1.1.6: September 11th, 2018
 * **New** - Order package/accommodation/spatreatment/discovery by order_menu